backend/requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0
